+++ packages/fast-components-react-base/src/hypertext/hypertext.tsx
@@ -1,20 +1,20 @@
-import React from "react";
-import ReactDOM from "react-dom";
+import { HypertextClassNameContract } from "@microsoft/fast-components-class-name-contracts-base";
 import Foundation, { HandledProps } from "@microsoft/fast-components-foundation-react";
+import { get } from "lodash-es";
+import React from "react";
+import { DisplayNamePrefix } from "../utilities";
 import {
     HypertextHandledProps,
-    HypertextManagedClasses,
+    HypertextProps,
     HypertextUnhandledProps,
 } from "./hypertext.props";
-import {
-    HypertextClassNameContract,
-    ManagedClasses,
-} from "@microsoft/fast-components-class-name-contracts-base";
-import { get } from "lodash-es";
-import { DisplayNamePrefix } from "../utilities";
 
 class Hypertext extends Foundation<HypertextHandledProps, HypertextUnhandledProps, {}> {
     public static displayName: string = `${DisplayNamePrefix}Hypertext`;
+
+    public static defaultProps: Partial<HypertextProps> = {
+        managedClasses: {},
+    };
 
     protected handledProps: HandledProps<HypertextHandledProps> = {
         managedClasses: void 0,
@@ -39,7 +39,7 @@ class Hypertext extends Foundation<HypertextHandledProps, HypertextUnhandledProp
      * Generates class names
      */
     protected generateClassNames(): string {
-        return super.generateClassNames(get(this.props, "managedClasses.hypertext", ""));
+        return super.generateClassNames(this.props.managedClasses.hypertext);
     }
 }