+++ packages/fast-components-styles-msft/src/outline-button/index.ts
@@ -50,12 +50,12 @@ const styles: ComponentStyles<LightweightButtonClassNameContract, DesignSystem> 
             ),
         },
         ...applyFocusVisible<DesignSystem>({
-            boxShadow: format(
+            "box-shadow": format(
                 "0 0 0 {0} {1} inset",
                 toPx(subtract(focusOutlineWidth, outlineWidth)),
                 neutralFocus
             ),
-            borderColor: neutralFocus,
+            "border-color": neutralFocus,
             ...highContrastOutlineFocus,
         }),
         "&:disabled": {