+++ packages/fast-components-styles-msft/src/radio/index.ts
@@ -45,8 +45,8 @@ const styles: ComponentStyles<RadioClassNameContract, DesignSystem> = {
     radio: {
         position: "relative",
         display: "inline-flex",
-        flexDirection: "row",
-        alignItems: "center",
+        "flex-direction": "row",
+        "align-items": "center",
         transition: "all 0.2s ease-in-out",
     },
     radio_input: {
@@ -54,9 +54,11 @@ const styles: ComponentStyles<RadioClassNameContract, DesignSystem> = {
         width: inputSize,
         height: inputSize,
         appearance: "none",
-        borderRadius: "50%",
+        "-webkit-appearance": "none",
+        "-moz-appearance": "none",
+        "border-radius": "50%",
         margin: "0",
-        zIndex: "1",
+        "z-index": "1",
         background: neutralFillInputRest,
         transition: "all 0.2s ease-in-out",
         border: format(
@@ -69,17 +71,17 @@ const styles: ComponentStyles<RadioClassNameContract, DesignSystem> = {
         },
         "&:hover:enabled": {
             background: neutralFillInputHover,
-            borderColor: neutralOutlineHover,
+            "border-color": neutralOutlineHover,
         },
         "&:active": {
             background: neutralFillInputActive,
-            borderColor: neutralOutlineActive,
+            "border-color": neutralOutlineActive,
         },
         ...applyFocusVisible({
-            boxShadow: format<DesignSystem>("0 0 0 1px {0} inset", neutralFocus),
-            borderColor: neutralFocus,
+            "box-shadow": format<DesignSystem>("0 0 0 1px {0} inset", neutralFocus),
+            "border-color": neutralFocus,
             [highContrastSelector]: {
-                boxShadow: "0 0 0 1px ButtonText inset",
+                "box-shadow": "0 0 0 1px ButtonText inset",
             },
         }),
         [highContrastSelector]: {
@@ -88,17 +90,17 @@ const styles: ComponentStyles<RadioClassNameContract, DesignSystem> = {
     },
     radio_stateIndicator: {
         position: "relative",
-        borderRadius: "50%",
+        "border-radius": "50%",
         display: "inline-block",
         width: inputSize,
         height: inputSize,
-        flexShrink: "0",
+        "flex-shrink": "0",
         "&::before": {
-            pointerEvents: "none",
+            "pointer-events": "none",
             position: "absolute",
-            zIndex: "1",
+            "z-index": "1",
             content: '""',
-            borderRadius: "50%",
+            "border-radius": "50%",
             top: indicatorMargin,
             left: indicatorMargin,
             bottom: indicatorMargin,
@@ -110,8 +112,8 @@ const styles: ComponentStyles<RadioClassNameContract, DesignSystem> = {
         ...applyCursorPointer(),
         color: neutralForegroundRest,
         ...applyScaledTypeRamp("t7"),
-        marginLeft: directionSwitch(horizontalSpacing(2), ""),
-        marginRight: directionSwitch("", horizontalSpacing(2)),
+        "margin-left": directionSwitch(horizontalSpacing(2), ""),
+        "margin-right": directionSwitch("", horizontalSpacing(2)),
         [highContrastSelector]: {
             color: "ButtonText",
         },