+++ packages/fast-jss-utilities/src/truncation.ts
@@ -1,15 +1,19 @@
 import * as CSS from "csstype";
+import { CSSRules } from "@microsoft/fast-jss-manager";
 
+/**
+ * @deprecated
+ */
 export interface Ellipsis {
     whiteSpace: CSS.WhiteSpaceProperty;
     overflow: CSS.OverflowProperty;
     textOverflow: CSS.TextOverflowProperty;
 }
 
-export function ellipsis(): Ellipsis {
+export function ellipsis(): CSSRules<any> {
     return {
-        whiteSpace: "nowrap",
+        "white-space": "nowrap",
         overflow: "hidden",
-        textOverflow: "ellipsis",
+        "text-overflow": "ellipsis",
     };
 }