+++ packages/fast-web-utilities/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.3.0](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-web-utilities@4.2.1...@microsoft/fast-web-utilities@4.3.0) (2019-08-22)
+
+
+### Features
+
+* adds classNames utility ([#2163](https://github.com/Microsoft/fast-dna/issues/2163)) ([d6a872d](https://github.com/Microsoft/fast-dna/commit/d6a872d))
+* export individual keycodes as named exports and unreference KeyCodes ([327d806](https://github.com/Microsoft/fast-dna/commit/327d806))
+
+
+
+
+
 ## [4.2.1](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-web-utilities@4.2.0...@microsoft/fast-web-utilities@4.2.1) (2019-08-09)